refactor: update command
